Revert "clean up java code."
This reverts commit 47e308d957793316707ddea930d543016338f69c.

+++ util/src/main/java/org/xipki/util/cbor/CborDecoder.java
@@ -777,7 +777,7 @@ public class CborDecoder implements AutoCloseable {
         }
 
         BigInteger[] ret = new BigInteger[arrayLen];
-        for (int i = 0; i < arrayLen; i++) {
+        for (int i = 0; i < arrayLen.intValue(); i++) {
             ret[i] = readBigInt();
         }
